+++ scripts/mdx-transforms/paragraph-custom-alert/__fixtures__/basic/test-file-results.mdx
@@ -1,0 +1,29 @@
+# Hello world!
+
+This is a test file. 
+
+<div className="alert alert-success g-type-body">
+
+This is a green success alert
+
+</div>
+
+<div className="alert alert-info g-type-body">
+
+This is a blue info alert
+
+</div>
+
+<div className="alert alert-warning g-type-body">
+
+This is a yellow warning alert
+
+</div>
+
+<div className="alert alert-danger g-type-body">
+
+This is a red danger alert
+
+</div>
+
+This is the end of the test file.
